feat - move class, ExchangeRate class to entity package
