+++ src/main/java/co/invest72/exchange_rate/domain/ExchangeRateRepository.java
@@ -1,5 +1,7 @@
 package co.invest72.exchange_rate.domain;
 
+import co.invest72.exchange_rate.domain.entity.ExchangeRate;
+
 public interface ExchangeRateRepository {
 	void save(ExchangeRate exchangeRate);
 }
