+++ src/main/java/co/invest72/exchange_rate/infrastructure/persistence/InMemoryExchangeRateRepository.java
@@ -3,7 +3,7 @@ package co.invest72.exchange_rate.infrastructure.persistence;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
-import co.invest72.exchange_rate.domain.ExchangeRate;
+import co.invest72.exchange_rate.domain.entity.ExchangeRate;
 import co.invest72.exchange_rate.domain.ExchangeRateRepository;
 
 public class InMemoryExchangeRateRepository implements ExchangeRateRepository {
